essaiCo.py
from comer.datamodule import vocab
from comer.lit_comer import LitCoMER
from torchvision.transforms import ToTensor
import torch
from PIL import Image

ckpt = 'lightning_logs/version_28/checkpoints/epoch=164-step=111209-val_ExpRate=0.7792.ckpt'
model = LitCoMER.load_from_checkpoint(ckpt)
model = model.eval()
device = torch.device("cpu")
model = model.to(device)

img = Image.open('crop.png')

img = img.convert("L")

img = ToTensor()(img)
mask = torch.zeros_like(img, dtype=torch.bool)
hyp = model.approximate_joint_search(img.unsqueeze(0), mask)[0]
pred_latex = vocab.indices2label(hyp.seq)

print(pred_latex)