requirements.txt
python=3.7
opencv-python==4.5.5.64
pytorch==1.12.1 
torchvision==0.13.1 
torchaudio==0.12.1 
pyqt=5
pytorch-lightning==1.4.9
einops==0.6.1 
torchmetrics==0.6.0
